backend/requirements.txt
djangorestframework
transformers
torch